+++ src/App.tsx
@@ -3,13 +3,13 @@ import { QueryClient, QueryClientProvider } from 'react-query';
 import { ReactQueryDevtools } from 'react-query/devtools';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import './App.css';
-import BenchmarkDetail from './components/benchmarks/BenchmarkDetail';
-import { Benchmarks } from './components/benchmarks/Benchmarks';
-import { CreateBenchmark } from './components/benchmarks/CreateBenchmark';
-import Dashboard from './components/Dashboard';
+import BenchmarkDetail from './components/Benchmarks/BenchmarkDetail';
+import { Benchmarks } from './components/Benchmarks/Benchmarks';
+import { CreateBenchmark } from './components/Benchmarks/CreateBenchmark';
+import Dashboard from './components/Dashboard/Dashboard';
 import Landing from './components/Landing';
 import Login from './components/Login';
-import PrivateRoute from './components/PrivateRoute';
+import PrivateRoute from './components/Routing/PrivateRoute';
 
 function App() {
   const queryClient = new QueryClient();
