+++ src/components/Login.tsx
@@ -2,7 +2,7 @@ import { LockClosedIcon } from '@heroicons/react/solid';
 import React from 'react';
 import { useHistory } from 'react-router-dom';
 import login from '../api/auth';
-import useToken from '../utils/useToken';
+import useToken from '../hooks/token';
 
 const Login: React.FC = () => {
   const history = useHistory();
